+++ eras/allegra/impl/CHANGELOG.md
@@ -11,6 +11,12 @@
 
 * Change `ipv4` and `ipv6` to use exact byte sizes (4 and 16 respectively), no longer allowing leftover bytes
 
+### `testlib`
+
+* In `Test.Cardano.Ledger.Allegra.Examples`:
+  - Remove `mkAllegraBasedExampleTx`, `exampleAllegraBasedTxBody`, `exampleAllegraBasedShelleyTxBody`
+  - Add `exampleAllegraBasedTx`
+
 ## 1.9.0.0
 
 * Add `Generic` instance for `ApplyTxError`