+++ eras/babbage/impl/CHANGELOG.md
@@ -28,6 +28,9 @@
 ### `testlib`
 
 * Add `Test.Cardano.Ledger.Babbage.Imp.PoolSpec` module with `babbageEraSpecificSpec`
+* In `Test.Cardano.Ledger.Babbage.Examples`:
+  - Remove `mkBabbageBasedExampleTx`, `exampleBabbageBasedTxBody`
+  - Add `exampleBabbageBasedTx`, `exampleBabbageBasedTopTx`
 
 ## 1.13.0.0